Moved exported methods out of test file
