+++ go/libraries/doltcore/sqle/enginetest/dolt_engine_test.go
@@ -17,7 +17,6 @@ package enginetest
 import (
 	"context"
 	"fmt"
-	"io"
 	"os"
 	"runtime"
 	"sync"
@@ -27,46 +26,25 @@ import (
 	"github.com/dolthub/go-mysql-server/enginetest"
 	"github.com/dolthub/go-mysql-server/enginetest/queries"
 	"github.com/dolthub/go-mysql-server/enginetest/scriptgen/setup"
-	"github.com/dolthub/go-mysql-server/server"
 	"github.com/dolthub/go-mysql-server/sql"
-	"github.com/dolthub/go-mysql-server/sql/memo"
 	"github.com/dolthub/go-mysql-server/sql/mysql_db"
 	"github.com/dolthub/go-mysql-server/sql/plan"
-	"github.com/dolthub/go-mysql-server/sql/transform"
 	gmstypes "github.com/dolthub/go-mysql-server/sql/types"
-	"github.com/dolthub/vitess/go/mysql"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb"
 	"github.com/dolthub/dolt/go/libraries/doltcore/dtestutils"
 	"github.com/dolthub/dolt/go/libraries/doltcore/env"
-	"github.com/dolthub/dolt/go/libraries/doltcore/schema"
 	"github.com/dolthub/dolt/go/libraries/doltcore/sqle"
 	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/dsess"
 	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/statspro"
 	"github.com/dolthub/dolt/go/libraries/utils/config"
-	"github.com/dolthub/dolt/go/store/datas"
 	"github.com/dolthub/dolt/go/store/types"
 )
-
-var skipPrepared bool
 
 // SkipPreparedsCount is used by the "ci-check-repo CI workflow
 // as a reminder to consider prepareds when adding a new
 // enginetest suite.
 const SkipPreparedsCount = 83
-
-const skipPreparedFlag = "DOLT_SKIP_PREPARED_ENGINETESTS"
-
-func init() {
-	sqle.MinRowsPerPartition = 8
-	sqle.MaxRowsPerPartition = 1024
-
-	if v := os.Getenv(skipPreparedFlag); v != "" {
-		skipPrepared = true
-	}
-}
 
 func TestQueries(t *testing.T) {
 	h := newDoltHarness(t)
@@ -124,27 +102,6 @@ func TestSchemaOverrides(t *testing.T) {
 	RunSchemaOverridesTest(t, harness)
 }
 
-func RunSchemaOverridesTest(t *testing.T, harness DoltEnginetestHarness) {
-	tcc := &testCommitClock{}
-	cleanup := installTestCommitClock(tcc)
-	defer cleanup()
-
-	for _, script := range SchemaOverrideTests {
-		sql.RunWithNowFunc(tcc.Now, func() error {
-			harness = harness.NewHarness(t)
-			harness.Setup(setup.MydbData)
-
-			engine, err := harness.NewEngine(t)
-			if err != nil {
-				panic(err)
-			}
-
-			enginetest.TestScriptWithEngine(t, engine, harness, script)
-			return nil
-		})
-	}
-}
-
 // Convenience test for debugging a single query. Unskip and set to the desired query.
 func TestSingleScript(t *testing.T) {
 	t.Skip()
@@ -182,124 +139,6 @@ func newUpdateResult(matched, updated int) gmstypes.OkResult {
 func TestAutoIncrementTrackerLockMode(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunAutoIncrementTrackerLockModeTest(t, harness)
-}
-
-func RunAutoIncrementTrackerLockModeTest(t *testing.T, harness DoltEnginetestHarness) {
-	for _, lockMode := range []int64{0, 1, 2} {
-		t.Run(fmt.Sprintf("lock mode %d", lockMode), func(t *testing.T) {
-			testAutoIncrementTrackerWithLockMode(t, harness, lockMode)
-		})
-	}
-}
-
-// testAutoIncrementTrackerWithLockMode tests that interleaved inserts don't cause deadlocks, regardless of the value of innodb_autoinc_lock_mode.
-// In a real use case, these interleaved operations would be happening in different sessions on different threads.
-// In order to make the test behave predictably, we manually interleave the two iterators.
-func testAutoIncrementTrackerWithLockMode(t *testing.T, harness DoltEnginetestHarness, lockMode int64) {
-	err := sql.SystemVariables.AssignValues(map[string]interface{}{"innodb_autoinc_lock_mode": lockMode})
-	require.NoError(t, err)
-
-	setupScripts := []setup.SetupScript{[]string{
-		"CREATE TABLE test1 (pk int NOT NULL PRIMARY KEY AUTO_INCREMENT,c0 int,index t1_c_index (c0));",
-		"CREATE TABLE test2 (pk int NOT NULL PRIMARY KEY AUTO_INCREMENT,c0 int,index t2_c_index (c0));",
-		"CREATE TABLE timestamps (pk int NOT NULL PRIMARY KEY AUTO_INCREMENT, t int);",
-		"CREATE TRIGGER t1 AFTER INSERT ON test1 FOR EACH ROW INSERT INTO timestamps VALUES (0, 1);",
-		"CREATE TRIGGER t2 AFTER INSERT ON test2 FOR EACH ROW INSERT INTO timestamps VALUES (0, 2);",
-		"CREATE VIEW bin AS SELECT 0 AS v UNION ALL SELECT 1;",
-		"CREATE VIEW sequence5bit AS SELECT b1.v + 2*b2.v + 4*b3.v + 8*b4.v + 16*b5.v AS v from bin b1, bin b2, bin b3, bin b4, bin b5;",
-	}}
-
-	harness = harness.NewHarness(t)
-	defer harness.Close()
-	harness.Setup(setup.MydbData, setupScripts)
-	e := mustNewEngine(t, harness)
-
-	defer e.Close()
-	ctx := enginetest.NewContext(harness)
-
-	// Confirm that the system variable was correctly set.
-	_, iter, err := e.Query(ctx, "select @@innodb_autoinc_lock_mode")
-	require.NoError(t, err)
-	rows, err := sql.RowIterToRows(ctx, iter)
-	require.NoError(t, err)
-	assert.Equal(t, rows, []sql.Row{{lockMode}})
-
-	// Ordinarily QueryEngine.query manages transactions.
-	// Since we can't use that for this test, we manually start a new transaction.
-	ts := ctx.Session.(sql.TransactionSession)
-	tx, err := ts.StartTransaction(ctx, sql.ReadWrite)
-	require.NoError(t, err)
-	ctx.SetTransaction(tx)
-
-	getTriggerIter := func(query string) sql.RowIter {
-		root, err := e.AnalyzeQuery(ctx, query)
-		require.NoError(t, err)
-
-		var triggerNode *plan.TriggerExecutor
-		transform.Node(root, func(n sql.Node) (sql.Node, transform.TreeIdentity, error) {
-			if triggerNode != nil {
-				return n, transform.SameTree, nil
-			}
-			if t, ok := n.(*plan.TriggerExecutor); ok {
-				triggerNode = t
-			}
-			return n, transform.NewTree, nil
-		})
-		iter, err := e.EngineAnalyzer().ExecBuilder.Build(ctx, triggerNode, nil)
-		require.NoError(t, err)
-		return iter
-	}
-
-	iter1 := getTriggerIter("INSERT INTO test1 (c0) select v from sequence5bit;")
-	iter2 := getTriggerIter("INSERT INTO test2 (c0) select v from sequence5bit;")
-
-	// Alternate the iterators until they're exhausted.
-	var err1 error
-	var err2 error
-	for err1 != io.EOF || err2 != io.EOF {
-		if err1 != io.EOF {
-			var row1 sql.Row
-			require.NoError(t, err1)
-			row1, err1 = iter1.Next(ctx)
-			_ = row1
-		}
-		if err2 != io.EOF {
-			require.NoError(t, err2)
-			_, err2 = iter2.Next(ctx)
-		}
-	}
-	err = iter1.Close(ctx)
-	require.NoError(t, err)
-	err = iter2.Close(ctx)
-	require.NoError(t, err)
-
-	dsess.DSessFromSess(ctx.Session).CommitTransaction(ctx, ctx.GetTransaction())
-
-	// Verify that the inserts are seen by the engine.
-	{
-		_, iter, err := e.Query(ctx, "select count(*) from timestamps")
-		require.NoError(t, err)
-		rows, err := sql.RowIterToRows(ctx, iter)
-		require.NoError(t, err)
-		assert.Equal(t, rows, []sql.Row{{int64(64)}})
-	}
-
-	// Verify that the insert operations are actually interleaved by inspecting the order that values were added to `timestamps`
-	{
-		_, iter, err := e.Query(ctx, "select (select min(pk) from timestamps where t = 1) < (select max(pk) from timestamps where t = 2)")
-		require.NoError(t, err)
-		rows, err := sql.RowIterToRows(ctx, iter)
-		require.NoError(t, err)
-		assert.Equal(t, rows, []sql.Row{{true}})
-	}
-
-	{
-		_, iter, err := e.Query(ctx, "select (select min(pk) from timestamps where t = 2) < (select max(pk) from timestamps where t = 1)")
-		require.NoError(t, err)
-		rows, err := sql.RowIterToRows(ctx, iter)
-		require.NoError(t, err)
-		assert.Equal(t, rows, []sql.Row{{true}})
-	}
 }
 
 // Convenience test for debugging a single query. Unskip and set to the desired query.
@@ -543,21 +382,6 @@ func TestVersionedQueries(t *testing.T) {
 	RunVersionedQueriesTest(t, h)
 }
 
-func RunVersionedQueriesTest(t *testing.T, h DoltEnginetestHarness) {
-	h.Setup(setup.MydbData, []setup.SetupScript{VersionedQuerySetup, VersionedQueryViews})
-
-	e, err := h.NewEngine(t)
-	require.NoError(t, err)
-
-	for _, tt := range queries.VersionedQueries {
-		enginetest.TestQueryWithEngine(t, h, e, tt)
-	}
-
-	for _, tt := range queries.VersionedScripts {
-		enginetest.TestScriptWithEngine(t, e, h, tt)
-	}
-}
-
 func TestAnsiQuotesSqlMode(t *testing.T) {
 	enginetest.TestAnsiQuotesSqlMode(t, newDoltHarness(t))
 }
@@ -571,44 +395,6 @@ func TestAnsiQuotesSqlModePrepared(t *testing.T) {
 func TestQueryPlans(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunQueryTestPlans(t, harness)
-}
-
-func RunQueryTestPlans(t *testing.T, harness DoltEnginetestHarness) {
-	// Dolt supports partial keys, so the index matched is different for some plans
-	// TODO: Fix these differences by implementing partial key matching in the memory tables, or the engine itself
-	skipped := []string{
-		"SELECT pk,pk1,pk2 FROM one_pk LEFT JOIN two_pk ON pk=pk1",
-		"SELECT pk,pk1,pk2 FROM one_pk JOIN two_pk ON pk=pk1",
-		"SELECT one_pk.c5,pk1,pk2 FROM one_pk JOIN two_pk ON pk=pk1 ORDER BY 1,2,3",
-		"SELECT opk.c5,pk1,pk2 FROM one_pk opk JOIN two_pk tpk ON opk.pk=tpk.pk1 ORDER BY 1,2,3",
-		"SELECT opk.c5,pk1,pk2 FROM one_pk opk JOIN two_pk tpk ON pk=pk1 ORDER BY 1,2,3",
-		"SELECT pk,pk1,pk2 FROM one_pk LEFT JOIN two_pk ON pk=pk1 ORDER BY 1,2,3",
-		"SELECT pk,pk1,pk2 FROM one_pk t1, two_pk t2 WHERE pk=1 AND pk2=1 AND pk1=1 ORDER BY 1,2",
-	}
-	// Parallelism introduces Exchange nodes into the query plans, so disable.
-	// TODO: exchange nodes should really only be part of the explain plan under certain debug settings
-	harness = harness.NewHarness(t).WithSkippedQueries(skipped).WithConfigureStats(true)
-	if !types.IsFormat_DOLT(types.Format_Default) {
-		// only new format supports reverse IndexTableAccess
-		reverseIndexSkip := []string{
-			"SELECT * FROM one_pk ORDER BY pk",
-			"SELECT * FROM two_pk ORDER BY pk1, pk2",
-			"SELECT * FROM two_pk ORDER BY pk1",
-			"SELECT pk1 AS one, pk2 AS two FROM two_pk ORDER BY pk1, pk2",
-			"SELECT pk1 AS one, pk2 AS two FROM two_pk ORDER BY one, two",
-			"SELECT i FROM (SELECT i FROM mytable ORDER BY i DESC LIMIT 1) sq WHERE i = 3",
-			"SELECT i FROM (SELECT i FROM (SELECT i FROM mytable ORDER BY DES LIMIT 1) sql1)sql2 WHERE i = 3",
-			"SELECT s,i FROM mytable order by i DESC",
-			"SELECT s,i FROM mytable as a order by i DESC",
-			"SELECT pk1, pk2 FROM two_pk order by pk1 asc, pk2 asc",
-			"SELECT pk1, pk2 FROM two_pk order by pk1 desc, pk2 desc",
-			"SELECT i FROM (SELECT i FROM (SELECT i FROM mytable ORDER BY i DESC  LIMIT 1) sq1) sq2 WHERE i = 3",
-		}
-		harness = harness.WithSkippedQueries(reverseIndexSkip)
-	}
-
-	defer harness.Close()
-	enginetest.TestQueryPlans(t, harness, queries.PlanTests)
 }
 
 func TestIntegrationQueryPlans(t *testing.T) {
@@ -626,43 +412,9 @@ func TestDoltDiffQueryPlans(t *testing.T) {
 	RunDoltDiffQueryPlansTest(t, harness)
 }
 
-func RunDoltDiffQueryPlansTest(t *testing.T, harness DoltEnginetestHarness) {
-	defer harness.Close()
-	harness.Setup(setup.SimpleSetup...)
-	e, err := harness.NewEngine(t)
-	require.NoError(t, err)
-	defer e.Close()
-
-	for _, tt := range append(DoltDiffPlanTests, DoltCommitPlanTests...) {
-		enginetest.TestQueryPlanWithName(t, tt.Query, harness, e, tt.Query, tt.ExpectedPlan, sql.DescribeOptions{})
-	}
-}
-
 func TestBranchPlans(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunBranchPlanTests(t, harness)
-}
-
-func RunBranchPlanTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, script := range BranchPlanTests {
-		t.Run(script.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-
-			e := mustNewEngine(t, harness)
-			defer e.Close()
-
-			for _, statement := range script.SetUpScript {
-				ctx := enginetest.NewContext(harness).WithQuery(statement)
-				enginetest.RunQueryWithContext(t, e, harness, ctx, statement)
-			}
-			for _, tt := range script.Queries {
-				t.Run(tt.Query, func(t *testing.T) {
-					TestIndexedAccess(t, e, harness, tt.Query, tt.Index)
-				})
-			}
-		})
-	}
 }
 
 func TestQueryErrors(t *testing.T) {
@@ -674,19 +426,6 @@ func TestQueryErrors(t *testing.T) {
 func TestInfoSchema(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunInfoSchemaTests(t, h)
-}
-
-func RunInfoSchemaTests(t *testing.T, h DoltEnginetestHarness) {
-	defer h.Close()
-	enginetest.TestInfoSchema(t, h)
-
-	for _, script := range DoltInfoSchemaScripts {
-		func() {
-			h = h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
 }
 
 func TestColumnAliases(t *testing.T) {
@@ -758,33 +497,9 @@ func TestInsertIntoErrors(t *testing.T) {
 	RunInsertIntoErrorsTest(t, h)
 }
 
-func RunInsertIntoErrorsTest(t *testing.T, h DoltEnginetestHarness) {
-	h = h.WithSkippedQueries([]string{
-		"create table bad (vb varbinary(65535))",
-		"insert into bad values (repeat('0', 65536))",
-	})
-	defer h.Close()
-	enginetest.TestInsertIntoErrors(t, h)
-}
-
 func TestGeneratedColumns(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunGeneratedColumnTests(t, harness)
-}
-
-func RunGeneratedColumnTests(t *testing.T, harness DoltEnginetestHarness) {
-	defer harness.Close()
-	enginetest.TestGeneratedColumns(t,
-		// virtual indexes are failing for certain lookups on this test
-		harness.WithSkippedQueries([]string{"create table t (pk int primary key, col1 int as (pk + 1));"}))
-
-	for _, script := range GeneratedColumnMergeTestScripts {
-		func() {
-			h := harness.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
 }
 
 func TestGeneratedColumnPlans(t *testing.T) {
@@ -1066,29 +781,9 @@ func TestBranchDdl(t *testing.T) {
 	RunBranchDdlTest(t, h)
 }
 
-func RunBranchDdlTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DdlBranchTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestBranchDdlPrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunBranchDdlTestPrepared(t, h)
-}
-
-func RunBranchDdlTestPrepared(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DdlBranchTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScriptPrepared(t, h, script)
-		}()
-	}
 }
 
 func TestPkOrdinalsDDL(t *testing.T) {
@@ -1164,14 +859,6 @@ func TestIndexPrefix(t *testing.T) {
 	RunIndexPrefixTest(t, harness)
 }
 
-func RunIndexPrefixTest(t *testing.T, harness *DoltHarness) {
-	defer harness.Close()
-	enginetest.TestIndexPrefix(t, harness)
-	for _, script := range DoltIndexPrefixScripts {
-		enginetest.TestScript(t, harness, script)
-	}
-}
-
 func TestBigBlobs(t *testing.T) {
 	skipOldFormat(t)
 
@@ -1179,58 +866,9 @@ func TestBigBlobs(t *testing.T) {
 	RunBigBlobsTest(t, h)
 }
 
-func RunBigBlobsTest(t *testing.T, h *DoltHarness) {
-	defer h.Close()
-	h.Setup(setup.MydbData, setup.BlobData)
-	for _, tt := range BigBlobQueries {
-		enginetest.RunWriteQueryTest(t, h, tt)
-	}
-}
-
 func TestDropDatabase(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDropEngineTest(t, h)
-}
-
-func RunDropEngineTest(t *testing.T, h DoltEnginetestHarness) {
-	func() {
-		h := h.NewHarness(t)
-		defer h.Close()
-		enginetest.TestScript(t, h, queries.ScriptTest{
-			Name: "Drop database engine tests for Dolt only",
-			SetUpScript: []string{
-				"CREATE DATABASE Test1db",
-				"CREATE DATABASE TEST2db",
-			},
-			Assertions: []queries.ScriptTestAssertion{
-				{
-					Query:    "DROP DATABASE TeSt2DB",
-					Expected: []sql.Row{{gmstypes.OkResult{RowsAffected: 1}}},
-				},
-				{
-					Query:       "USE test2db",
-					ExpectedErr: sql.ErrDatabaseNotFound,
-				},
-				{
-					Query:    "USE TEST1DB",
-					Expected: []sql.Row{},
-				},
-				{
-					Query:    "DROP DATABASE IF EXISTS test1DB",
-					Expected: []sql.Row{{gmstypes.OkResult{RowsAffected: 1}}},
-				},
-				{
-					Query:       "USE Test1db",
-					ExpectedErr: sql.ErrDatabaseNotFound,
-				},
-			},
-		})
-	}()
-
-	t.Skip("Dolt doesn't yet support dropping the primary database, which these tests do")
-	h = h.NewHarness(t)
-	defer h.Close()
-	enginetest.TestDropDatabase(t, h)
 }
 
 func TestCreateForeignKeys(t *testing.T) {
@@ -1256,67 +894,9 @@ func TestForeignKeyBranches(t *testing.T) {
 	RunForeignKeyBranchesTest(t, h)
 }
 
-func RunForeignKeyBranchesTest(t *testing.T, h DoltEnginetestHarness) {
-	setupPrefix := []string{
-		"call dolt_branch('b1')",
-		"use mydb/b1",
-	}
-	assertionsPrefix := []queries.ScriptTestAssertion{
-		{
-			Query:            "use mydb/b1",
-			SkipResultsCheck: true,
-		},
-	}
-	for _, script := range queries.ForeignKeyTests {
-		// New harness for every script because we create branches
-		h := h.NewHarness(t)
-		h.Setup(setup.MydbData, setup.Parent_childData)
-		modifiedScript := script
-		modifiedScript.SetUpScript = append(setupPrefix, modifiedScript.SetUpScript...)
-		modifiedScript.Assertions = append(assertionsPrefix, modifiedScript.Assertions...)
-		enginetest.TestScript(t, h, modifiedScript)
-	}
-
-	for _, script := range ForeignKeyBranchTests {
-		// New harness for every script because we create branches
-		h := h.NewHarness(t)
-		h.Setup(setup.MydbData, setup.Parent_childData)
-		enginetest.TestScript(t, h, script)
-	}
-}
-
 func TestForeignKeyBranchesPrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunForeignKeyBranchesPreparedTest(t, h)
-}
-
-func RunForeignKeyBranchesPreparedTest(t *testing.T, h DoltEnginetestHarness) {
-	setupPrefix := []string{
-		"call dolt_branch('b1')",
-		"use mydb/b1",
-	}
-	assertionsPrefix := []queries.ScriptTestAssertion{
-		{
-			Query:            "use mydb/b1",
-			SkipResultsCheck: true,
-		},
-	}
-	for _, script := range queries.ForeignKeyTests {
-		// New harness for every script because we create branches
-		h := h.NewHarness(t)
-		h.Setup(setup.MydbData, setup.Parent_childData)
-		modifiedScript := script
-		modifiedScript.SetUpScript = append(setupPrefix, modifiedScript.SetUpScript...)
-		modifiedScript.Assertions = append(assertionsPrefix, modifiedScript.Assertions...)
-		enginetest.TestScriptPrepared(t, h, modifiedScript)
-	}
-
-	for _, script := range ForeignKeyBranchTests {
-		// New harness for every script because we create branches
-		h := h.NewHarness(t)
-		h.Setup(setup.MydbData, setup.Parent_childData)
-		enginetest.TestScriptPrepared(t, h, script)
-	}
 }
 
 func TestFulltextIndexes(t *testing.T) {
@@ -1378,49 +958,14 @@ func TestBranchViews(t *testing.T) {
 	RunBranchViewsTest(t, h)
 }
 
-func RunBranchViewsTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range ViewBranchTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestBranchViewsPrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunBranchViewsPreparedTest(t, h)
 }
 
-func RunBranchViewsPreparedTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range ViewBranchTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScriptPrepared(t, h, script)
-		}()
-	}
-}
-
 func TestVersionedViews(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunVersionedViewsTest(t, h)
-}
-
-func RunVersionedViewsTest(t *testing.T, h DoltEnginetestHarness) {
-	defer h.Close()
-	h.Setup(setup.MydbData, []setup.SetupScript{VersionedQuerySetup, VersionedQueryViews})
-
-	e, err := h.NewEngine(t)
-	require.NoError(t, err)
-
-	for _, testCase := range queries.VersionedViewTests {
-		t.Run(testCase.Query, func(t *testing.T) {
-			ctx := enginetest.NewContext(h)
-			enginetest.TestQueryWithContext(t, ctx, e, h, testCase.Query, testCase.Expected, testCase.ExpectedColumns, nil)
-		})
-	}
 }
 
 func TestWindowFunctions(t *testing.T) {
@@ -1495,14 +1040,6 @@ func TestVariables(t *testing.T) {
 	RunVariableTest(t, h)
 }
 
-func RunVariableTest(t *testing.T, h DoltEnginetestHarness) {
-	defer h.Close()
-	enginetest.TestVariables(t, h)
-	for _, script := range DoltSystemVariables {
-		enginetest.TestScript(t, h, script)
-	}
-}
-
 func TestVariableErrors(t *testing.T) {
 	h := newDoltHarness(t)
 	defer h.Close()
@@ -1564,48 +1101,14 @@ func TestStoredProcedures(t *testing.T) {
 	RunStoredProceduresTest(t, h)
 }
 
-func RunStoredProceduresTest(t *testing.T, h DoltEnginetestHarness) {
-	tests := make([]queries.ScriptTest, 0, len(queries.ProcedureLogicTests))
-	for _, test := range queries.ProcedureLogicTests {
-		// TODO: this passes locally but SOMETIMES fails tests on GitHub, no clue why
-		if test.Name != "ITERATE and LEAVE loops" {
-			tests = append(tests, test)
-		}
-	}
-	queries.ProcedureLogicTests = tests
-
-	defer h.Close()
-	enginetest.TestStoredProcedures(t, h)
-}
-
 func TestDoltStoredProcedures(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltStoredProceduresTest(t, h)
 }
 
-func RunDoltStoredProceduresTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltProcedureTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestDoltStoredProceduresPrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltStoredProceduresPreparedTest(t, h)
-}
-
-func RunDoltStoredProceduresPreparedTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltProcedureTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScriptPrepared(t, h, script)
-		}()
-	}
 }
 
 func TestEvents(t *testing.T) {
@@ -1619,33 +1122,9 @@ func TestCallAsOf(t *testing.T) {
 	RunCallAsOfTest(t, h)
 }
 
-func RunCallAsOfTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltCallAsOf {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestLargeJsonObjects(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunLargeJsonObjectsTest(t, harness)
-}
-
-func RunLargeJsonObjectsTest(t *testing.T, harness DoltEnginetestHarness) {
-	SkipByDefaultInCI(t)
-	defer harness.Close()
-	for _, script := range LargeJsonObjectScriptTests {
-		enginetest.TestScript(t, harness, script)
-	}
-}
-
-func SkipByDefaultInCI(t *testing.T) {
-	if os.Getenv("CI") != "" && os.Getenv("DOLT_TEST_RUN_NON_RACE_TESTS") == "" {
-		t.Skip()
-	}
 }
 
 func TestTransactions(t *testing.T) {
@@ -1653,57 +1132,9 @@ func TestTransactions(t *testing.T) {
 	RunTransactionTests(t, h)
 }
 
-func RunTransactionTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range queries.TransactionTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestTransactionScript(t, h, script)
-		}()
-	}
-	for _, script := range DoltTransactionTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestTransactionScript(t, h, script)
-		}()
-	}
-	for _, script := range DoltStoredProcedureTransactionTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestTransactionScript(t, h, script)
-		}()
-	}
-	for _, script := range DoltConflictHandlingTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestTransactionScript(t, h, script)
-		}()
-	}
-	for _, script := range DoltConstraintViolationTransactionTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestTransactionScript(t, h, script)
-		}()
-	}
-}
-
 func TestBranchTransactions(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunBranchTransactionTest(t, h)
-}
-
-func RunBranchTransactionTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range BranchIsolationTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestTransactionScript(t, h, script)
-		}()
-	}
 }
 
 func TestMultiDbTransactions(t *testing.T) {
@@ -1711,37 +1142,9 @@ func TestMultiDbTransactions(t *testing.T) {
 	RunMultiDbTransactionsTest(t, h)
 }
 
-func RunMultiDbTransactionsTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range MultiDbTransactionTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-
-	for _, script := range MultiDbSavepointTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestTransactionScript(t, h, script)
-		}()
-	}
-}
-
 func TestMultiDbTransactionsPrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunMultiDbTransactionsPreparedTest(t, h)
-}
-
-func RunMultiDbTransactionsPreparedTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range MultiDbTransactionTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScriptPrepared(t, h, script)
-		}()
-	}
 }
 
 func TestConcurrentTransactions(t *testing.T) {
@@ -1755,27 +1158,9 @@ func TestDoltScripts(t *testing.T) {
 	RunDoltScriptsTest(t, harness)
 }
 
-func RunDoltScriptsTest(t *testing.T, harness DoltEnginetestHarness) {
-	for _, script := range DoltScripts {
-		go func() {
-			harness := harness.NewHarness(t)
-			defer harness.Close()
-			enginetest.TestScript(t, harness, script)
-		}()
-	}
-}
-
 func TestDoltTempTableScripts(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunDoltTempTableScripts(t, harness)
-}
-
-func RunDoltTempTableScripts(t *testing.T, harness DoltEnginetestHarness) {
-	for _, script := range DoltTempTableScripts {
-		harness := harness.NewHarness(t)
-		enginetest.TestScript(t, harness, script)
-		harness.Close()
-	}
 }
 
 func TestDoltRevisionDbScripts(t *testing.T) {
@@ -1783,162 +1168,14 @@ func TestDoltRevisionDbScripts(t *testing.T) {
 	RunDoltRevisionDbScriptsTest(t, h)
 }
 
-func RunDoltRevisionDbScriptsTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltRevisionDbScripts {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-
-	// Testing a commit-qualified database revision spec requires
-	// a little extra work to get the generated commit hash
-	h = h.NewHarness(t)
-	defer h.Close()
-	e, err := h.NewEngine(t)
-	require.NoError(t, err)
-	defer e.Close()
-	ctx := h.NewContext()
-
-	setupScripts := []setup.SetupScript{
-		{"create table t01 (pk int primary key, c1 int)"},
-		{"call dolt_add('.');"},
-		{"call dolt_commit('-am', 'creating table t01 on main');"},
-		{"insert into t01 values (1, 1), (2, 2);"},
-		{"call dolt_commit('-am', 'adding rows to table t01 on main');"},
-		{"insert into t01 values (3, 3);"},
-		{"call dolt_commit('-am', 'adding another row to table t01 on main');"},
-	}
-	_, err = enginetest.RunSetupScripts(ctx, h.Engine(), setupScripts, true)
-	require.NoError(t, err)
-
-	_, iter, err := h.Engine().Query(ctx, "select hashof('HEAD~2');")
-	require.NoError(t, err)
-	rows, err := sql.RowIterToRows(ctx, iter)
-	require.NoError(t, err)
-	assert.Equal(t, 1, len(rows))
-	commithash := rows[0][0].(string)
-
-	scriptTest := queries.ScriptTest{
-		Name: "database revision specs: commit-qualified revision spec",
-		Assertions: []queries.ScriptTestAssertion{
-			{
-				Query:    "show databases;",
-				Expected: []sql.Row{{"mydb"}, {"information_schema"}, {"mysql"}},
-			},
-			{
-				Query:    "use mydb/" + commithash,
-				Expected: []sql.Row{},
-			},
-			{
-				Query: "select active_branch();",
-				Expected: []sql.Row{
-					{nil},
-				},
-			},
-			{
-				Query:    "select database();",
-				Expected: []sql.Row{{"mydb/" + commithash}},
-			},
-			{
-				Query:    "show databases;",
-				Expected: []sql.Row{{"mydb"}, {"mydb/" + commithash}, {"information_schema"}, {"mysql"}},
-			},
-			{
-				Query:    "select * from t01",
-				Expected: []sql.Row{},
-			},
-			{
-				Query:          "call dolt_reset();",
-				ExpectedErrStr: "unable to reset HEAD in read-only databases",
-			},
-			{
-				Query:    "call dolt_checkout('main');",
-				Expected: []sql.Row{{0, "Switched to branch 'main'"}},
-			},
-			{
-				Query:    "select database();",
-				Expected: []sql.Row{{"mydb"}},
-			},
-			{
-				Query:    "select active_branch();",
-				Expected: []sql.Row{{"main"}},
-			},
-			{
-				Query:    "use mydb;",
-				Expected: []sql.Row{},
-			},
-			{
-				Query:    "select database();",
-				Expected: []sql.Row{{"mydb"}},
-			},
-			{
-				Query:    "show databases;",
-				Expected: []sql.Row{{"mydb"}, {"information_schema"}, {"mysql"}},
-			},
-		},
-	}
-
-	enginetest.TestScript(t, h, scriptTest)
-}
-
 func TestDoltRevisionDbScriptsPrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltRevisionDbScriptsPreparedTest(t, h)
 }
 
-func RunDoltRevisionDbScriptsPreparedTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltRevisionDbScripts {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScriptPrepared(t, h, script)
-		}()
-	}
-}
-
 func TestDoltDdlScripts(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunDoltDdlScripts(t, harness)
-}
-
-func RunDoltDdlScripts(t *testing.T, harness DoltEnginetestHarness) {
-	defer harness.Close()
-	harness.Setup()
-
-	for _, script := range ModifyAndChangeColumnScripts {
-		e, err := harness.NewEngine(t)
-		require.NoError(t, err)
-		enginetest.TestScriptWithEngine(t, e, harness, script)
-	}
-
-	for _, script := range ModifyColumnTypeScripts {
-		e, err := harness.NewEngine(t)
-		require.NoError(t, err)
-		enginetest.TestScriptWithEngine(t, e, harness, script)
-	}
-
-	for _, script := range DropColumnScripts {
-		e, err := harness.NewEngine(t)
-		require.NoError(t, err)
-		enginetest.TestScriptWithEngine(t, e, harness, script)
-	}
-	if !types.IsFormat_DOLT(types.Format_Default) {
-		t.Skip("not fixing unique index on keyless tables for old format")
-	}
-	for _, script := range AddIndexScripts {
-		e, err := harness.NewEngine(t)
-		require.NoError(t, err)
-		enginetest.TestScriptWithEngine(t, e, harness, script)
-	}
-
-	// TODO: these scripts should be general enough to go in GMS
-	for _, script := range AddDropPrimaryKeysScripts {
-		e, err := harness.NewEngine(t)
-		require.NoError(t, err)
-		enginetest.TestScriptWithEngine(t, e, harness, script)
-	}
 }
 
 func TestBrokenDdlScripts(t *testing.T) {
@@ -1958,29 +1195,9 @@ func TestShowCreateTable(t *testing.T) {
 	RunShowCreateTableTests(t, h)
 }
 
-func RunShowCreateTableTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range ShowCreateTableScriptTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestShowCreateTablePrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunShowCreateTablePreparedTests(t, h)
-}
-
-func RunShowCreateTablePreparedTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range ShowCreateTableScriptTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScriptPrepared(t, h, script)
-		}()
-	}
 }
 
 func TestViewsWithAsOf(t *testing.T) {
@@ -2001,34 +1218,9 @@ func TestDoltMerge(t *testing.T) {
 	RunDoltMergeTests(t, h)
 }
 
-func RunDoltMergeTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range MergeScripts {
-		// harness can't reset effectively when there are new commits / branches created, so use a new harness for
-		// each script
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			h.Setup(setup.MydbData)
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestDoltMergePrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltMergePreparedTests(t, h)
-}
-
-func RunDoltMergePreparedTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range MergeScripts {
-		// harness can't reset effectively when there are new commits / branches created, so use a new harness for
-		// each script
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScriptPrepared(t, h, script)
-		}()
-	}
 }
 
 func TestDoltRebase(t *testing.T) {
@@ -2036,33 +1228,9 @@ func TestDoltRebase(t *testing.T) {
 	RunDoltRebaseTests(t, h)
 }
 
-func RunDoltRebaseTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltRebaseScriptTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			h.SkipSetupCommit()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-
-	testMultiSessionScriptTests(t, DoltRebaseMultiSessionScriptTests)
-}
-
 func TestDoltRebasePrepared(t *testing.T) {
 	h := newDoltHarness(t)
 	RunDoltRebasePreparedTests(t, h)
-}
-
-func RunDoltRebasePreparedTests(t *testing.T, h *DoltHarness) {
-	for _, script := range DoltRebaseScriptTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			h.SkipSetupCommit()
-			enginetest.TestScriptPrepared(t, h, script)
-		}()
-	}
 }
 
 func TestDoltRevert(t *testing.T) {
@@ -2070,31 +1238,9 @@ func TestDoltRevert(t *testing.T) {
 	RunDoltRevertTests(t, h)
 }
 
-func RunDoltRevertTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range RevertScripts {
-		// harness can't reset effectively. Use a new harness for each script
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestDoltRevertPrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltRevertPreparedTests(t, h)
-}
-
-func RunDoltRevertPreparedTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range RevertScripts {
-		// harness can't reset effectively. Use a new harness for each script
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
 }
 
 func TestDoltAutoIncrement(t *testing.T) {
@@ -2102,56 +1248,14 @@ func TestDoltAutoIncrement(t *testing.T) {
 	RunDoltAutoIncrementTests(t, h)
 }
 
-func RunDoltAutoIncrementTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltAutoIncrementTests {
-		// doing commits on different branches is antagonistic to engine reuse, use a new engine on each script
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestDoltAutoIncrementPrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltAutoIncrementPreparedTests(t, h)
 }
 
-func RunDoltAutoIncrementPreparedTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltAutoIncrementTests {
-		// doing commits on different branches is antagonistic to engine reuse, use a new engine on each script
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestDoltConflictsTableNameTable(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltConflictsTableNameTableTests(t, h)
-}
-
-func RunDoltConflictsTableNameTableTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltConflictTableNameTableTests {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-
-	if types.IsFormat_DOLT(types.Format_Default) {
-		for _, script := range Dolt1ConflictTableNameTableTests {
-			func() {
-				h := h.NewHarness(t)
-				defer h.Close()
-				enginetest.TestScript(t, h, script)
-			}()
-		}
-	}
 }
 
 // tests new format behavior for keyless merges that create CVs and conflicts
@@ -2160,38 +1264,10 @@ func TestKeylessDoltMergeCVsAndConflicts(t *testing.T) {
 	RunKelyessDoltMergeCVsAndConflictsTests(t, h)
 }
 
-func RunKelyessDoltMergeCVsAndConflictsTests(t *testing.T, h DoltEnginetestHarness) {
-	if !types.IsFormat_DOLT(types.Format_Default) {
-		t.Skip()
-	}
-	for _, script := range KeylessMergeCVsAndConflictsScripts {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 // eventually this will be part of TestDoltMerge
 func TestDoltMergeArtifacts(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltMergeArtifacts(t, h)
-}
-
-func RunDoltMergeArtifacts(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range MergeArtifactsScripts {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-	for _, script := range SchemaConflictScripts {
-		h := h.NewHarness(t)
-		enginetest.TestScript(t, h, script)
-		h.Close()
-	}
 }
 
 // these tests are temporary while there is a difference between the old format
@@ -2214,17 +1290,6 @@ func TestDoltReset(t *testing.T) {
 	RunDoltResetTest(t, h)
 }
 
-func RunDoltResetTest(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltReset {
-		// dolt versioning conflicts with reset harness -- use new harness every time
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestDoltGC(t *testing.T) {
 	t.SkipNow()
 	for _, script := range DoltGC {
@@ -2241,51 +1306,9 @@ func TestDoltCheckout(t *testing.T) {
 	RunDoltCheckoutTests(t, h)
 }
 
-func RunDoltCheckoutTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltCheckoutScripts {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-
-	h = h.NewHarness(t)
-	defer h.Close()
-	engine, err := h.NewEngine(t)
-	require.NoError(t, err)
-	readOnlyEngine, err := h.NewReadOnlyEngine(engine.EngineAnalyzer().Catalog.DbProvider)
-	require.NoError(t, err)
-
-	for _, script := range DoltCheckoutReadOnlyScripts {
-		enginetest.TestScriptWithEngine(t, readOnlyEngine, h, script)
-	}
-}
-
 func TestDoltCheckoutPrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltCheckoutPreparedTests(t, h)
-}
-
-func RunDoltCheckoutPreparedTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltCheckoutScripts {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-
-	h = h.NewHarness(t)
-	defer h.Close()
-	engine, err := h.NewEngine(t)
-	require.NoError(t, err)
-	readOnlyEngine, err := h.NewReadOnlyEngine(engine.EngineAnalyzer().Catalog.DbProvider)
-	require.NoError(t, err)
-
-	for _, script := range DoltCheckoutReadOnlyScripts {
-		enginetest.TestScriptWithEnginePrepared(t, readOnlyEngine, h, script)
-	}
 }
 
 func TestDoltBranch(t *testing.T) {
@@ -2293,29 +1316,9 @@ func TestDoltBranch(t *testing.T) {
 	RunDoltBranchTests(t, h)
 }
 
-func RunDoltBranchTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltBranchScripts {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestDoltTag(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltTagTests(t, h)
-}
-
-func RunDoltTagTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltTagTestScripts {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
 }
 
 func TestDoltRemote(t *testing.T) {
@@ -2323,48 +1326,9 @@ func TestDoltRemote(t *testing.T) {
 	RunDoltRemoteTests(t, h)
 }
 
-func RunDoltRemoteTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltRemoteTestScripts {
-		func() {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 func TestDoltUndrop(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltUndropTests(t, h)
-}
-
-func RunDoltUndropTests(t *testing.T, h DoltEnginetestHarness) {
-	h.UseLocalFileSystem()
-	defer h.Close()
-	for _, script := range DoltUndropTestScripts {
-		enginetest.TestScript(t, h, script)
-	}
-}
-
-type testCommitClock struct {
-	unixNano int64
-}
-
-func (tcc *testCommitClock) Now() time.Time {
-	now := time.Unix(0, tcc.unixNano)
-	tcc.unixNano += int64(time.Hour)
-	return now
-}
-
-func installTestCommitClock(tcc *testCommitClock) func() {
-	oldNowFunc := datas.CommitterDate
-	oldCommitLoc := datas.CommitLoc
-	datas.CommitterDate = tcc.Now
-	datas.CommitLoc = time.UTC
-	return func() {
-		datas.CommitterDate = oldNowFunc
-		datas.CommitLoc = oldCommitLoc
-	}
 }
 
 // TestSingleTransactionScript is a convenience method for debugging a single transaction test. Unskip and set to the
@@ -2452,29 +1416,9 @@ func TestHistorySystemTable(t *testing.T) {
 	RunHistorySystemTableTests(t, harness)
 }
 
-func RunHistorySystemTableTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range HistorySystemTableScriptTests {
-		harness = harness.NewHarness(t)
-		harness.Setup(setup.MydbData)
-		t.Run(test.Name, func(t *testing.T) {
-			enginetest.TestScript(t, harness, test)
-		})
-	}
-}
-
 func TestHistorySystemTablePrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t).WithParallelism(2)
 	RunHistorySystemTableTestsPrepared(t, harness)
-}
-
-func RunHistorySystemTableTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range HistorySystemTableScriptTests {
-		harness = harness.NewHarness(t)
-		harness.Setup(setup.MydbData)
-		t.Run(test.Name, func(t *testing.T) {
-			enginetest.TestScriptPrepared(t, harness, test)
-		})
-	}
 }
 
 func TestBrokenHistorySystemTablePrepared(t *testing.T) {
@@ -2495,29 +1439,9 @@ func TestUnscopedDiffSystemTable(t *testing.T) {
 	RunUnscopedDiffSystemTableTests(t, h)
 }
 
-func RunUnscopedDiffSystemTableTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, test := range UnscopedDiffSystemTableScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScript(t, h, test)
-		})
-	}
-}
-
 func TestUnscopedDiffSystemTablePrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunUnscopedDiffSystemTableTestsPrepared(t, h)
-}
-
-func RunUnscopedDiffSystemTableTestsPrepared(t *testing.T, h DoltEnginetestHarness) {
-	for _, test := range UnscopedDiffSystemTableScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			h := h.NewHarness(t)
-			defer h.Close()
-			enginetest.TestScriptPrepared(t, h, test)
-		})
-	}
 }
 
 func TestColumnDiffSystemTable(t *testing.T) {
@@ -2525,31 +1449,9 @@ func TestColumnDiffSystemTable(t *testing.T) {
 	RunColumnDiffSystemTableTests(t, h)
 }
 
-func RunColumnDiffSystemTableTests(t *testing.T, h DoltEnginetestHarness) {
-	if !types.IsFormat_DOLT(types.Format_Default) {
-		t.Skip("correct behavior of dolt_column_diff only guaranteed on new format")
-	}
-	for _, test := range ColumnDiffSystemTableScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			enginetest.TestScript(t, h.NewHarness(t), test)
-		})
-	}
-}
-
 func TestColumnDiffSystemTablePrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunColumnDiffSystemTableTestsPrepared(t, h)
-}
-
-func RunColumnDiffSystemTableTestsPrepared(t *testing.T, h DoltEnginetestHarness) {
-	if !types.IsFormat_DOLT(types.Format_Default) {
-		t.Skip("correct behavior of dolt_column_diff only guaranteed on new format")
-	}
-	for _, test := range ColumnDiffSystemTableScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			enginetest.TestScriptPrepared(t, h.NewHarness(t), test)
-		})
-	}
 }
 
 func TestStatBranchTests(t *testing.T) {
@@ -2557,39 +1459,9 @@ func TestStatBranchTests(t *testing.T) {
 	RunStatBranchTests(t, harness)
 }
 
-func RunStatBranchTests(t *testing.T, harness DoltEnginetestHarness) {
-	defer harness.Close()
-	for _, test := range StatBranchTests {
-		t.Run(test.Name, func(t *testing.T) {
-			// reset engine so provider statistics are clean
-			harness = harness.NewHarness(t)
-			harness.Setup(setup.MydbData)
-			harness = harness.WithConfigureStats(true)
-			e := mustNewEngine(t, harness)
-			defer e.Close()
-			enginetest.TestScriptWithEngine(t, e, harness, test)
-		})
-	}
-}
-
 func TestStatsFunctions(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunStatsFunctionsTest(t, harness)
-}
-
-func RunStatsFunctionsTest(t *testing.T, harness DoltEnginetestHarness) {
-	defer harness.Close()
-	for _, test := range StatProcTests {
-		t.Run(test.Name, func(t *testing.T) {
-			// reset engine so provider statistics are clean
-			harness = harness.NewHarness(t).WithConfigureStats(true)
-			harness.Setup(setup.MydbData)
-			harness.SkipSetupCommit()
-			e := mustNewEngine(t, harness)
-			defer e.Close()
-			enginetest.TestScriptWithEngine(t, e, harness, test)
-		})
-	}
 }
 
 func TestDiffTableFunction(t *testing.T) {
@@ -2597,31 +1469,9 @@ func TestDiffTableFunction(t *testing.T) {
 	RunDiffTableFunctionTests(t, harness)
 }
 
-func RunDiffTableFunctionTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range DiffTableFunctionScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			enginetest.TestScript(t, harness, test)
-		})
-	}
-}
-
 func TestDiffTableFunctionPrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunDiffTableFunctionTestsPrepared(t, harness)
-}
-
-func RunDiffTableFunctionTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range DiffTableFunctionScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			enginetest.TestScriptPrepared(t, harness, test)
-		})
-	}
 }
 
 func TestDiffStatTableFunction(t *testing.T) {
@@ -2629,29 +1479,9 @@ func TestDiffStatTableFunction(t *testing.T) {
 	RunDiffStatTableFunctionTests(t, harness)
 }
 
-func RunDiffStatTableFunctionTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range DiffStatTableFunctionScriptTests {
-		harness = harness.NewHarness(t)
-		harness.Setup(setup.MydbData)
-		t.Run(test.Name, func(t *testing.T) {
-			enginetest.TestScript(t, harness, test)
-		})
-	}
-}
-
 func TestDiffStatTableFunctionPrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunDiffStatTableFunctionTestsPrepared(t, harness)
-}
-
-func RunDiffStatTableFunctionTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range DiffStatTableFunctionScriptTests {
-		harness = harness.NewHarness(t)
-		harness.Setup(setup.MydbData)
-		t.Run(test.Name, func(t *testing.T) {
-			enginetest.TestScriptPrepared(t, harness, test)
-		})
-	}
 }
 
 func TestDiffSummaryTableFunction(t *testing.T) {
@@ -2659,31 +1489,9 @@ func TestDiffSummaryTableFunction(t *testing.T) {
 	RunDiffSummaryTableFunctionTests(t, harness)
 }
 
-func RunDiffSummaryTableFunctionTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range DiffSummaryTableFunctionScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			enginetest.TestScript(t, harness, test)
-		})
-	}
-}
-
 func TestDiffSummaryTableFunctionPrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunDiffSummaryTableFunctionTestsPrepared(t, harness)
-}
-
-func RunDiffSummaryTableFunctionTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range DiffSummaryTableFunctionScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			enginetest.TestScriptPrepared(t, harness, test)
-		})
-	}
 }
 
 func TestPatchTableFunction(t *testing.T) {
@@ -2691,29 +1499,9 @@ func TestPatchTableFunction(t *testing.T) {
 	RunDoltPatchTableFunctionTests(t, harness)
 }
 
-func RunDoltPatchTableFunctionTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range PatchTableFunctionScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			harness.Setup(setup.MydbData)
-			enginetest.TestScript(t, harness, test)
-		})
-	}
-}
-
 func TestPatchTableFunctionPrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunDoltPatchTableFunctionTestsPrepared(t, harness)
-}
-
-func RunDoltPatchTableFunctionTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range PatchTableFunctionScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			harness.Setup(setup.MydbData)
-			enginetest.TestScriptPrepared(t, harness, test)
-		})
-	}
 }
 
 func TestLogTableFunction(t *testing.T) {
@@ -2721,33 +1509,9 @@ func TestLogTableFunction(t *testing.T) {
 	RunLogTableFunctionTests(t, harness)
 }
 
-func RunLogTableFunctionTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range LogTableFunctionScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			harness.SkipSetupCommit()
-			enginetest.TestScript(t, harness, test)
-		})
-	}
-}
-
 func TestLogTableFunctionPrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunLogTableFunctionTestsPrepared(t, harness)
-}
-
-func RunLogTableFunctionTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range LogTableFunctionScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			harness.SkipSetupCommit()
-			enginetest.TestScriptPrepared(t, harness, test)
-		})
-	}
 }
 
 func TestDoltReflog(t *testing.T) {
@@ -2773,31 +1537,9 @@ func TestCommitDiffSystemTable(t *testing.T) {
 	RunCommitDiffSystemTableTests(t, harness)
 }
 
-func RunCommitDiffSystemTableTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range CommitDiffSystemTableScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			enginetest.TestScript(t, harness, test)
-		})
-	}
-}
-
 func TestCommitDiffSystemTablePrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunCommitDiffSystemTableTestsPrepared(t, harness)
-}
-
-func RunCommitDiffSystemTableTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range CommitDiffSystemTableScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			enginetest.TestScriptPrepared(t, harness, test)
-		})
-	}
 }
 
 func TestDiffSystemTable(t *testing.T) {
@@ -2805,61 +1547,9 @@ func TestDiffSystemTable(t *testing.T) {
 	RunDoltDiffSystemTableTests(t, h)
 }
 
-func RunDoltDiffSystemTableTests(t *testing.T, h DoltEnginetestHarness) {
-	if !types.IsFormat_DOLT(types.Format_Default) {
-		t.Skip("only new format support system table indexing")
-	}
-
-	for _, test := range DiffSystemTableScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			h = h.NewHarness(t)
-			defer h.Close()
-			h.Setup(setup.MydbData)
-			enginetest.TestScript(t, h, test)
-		})
-	}
-
-	if types.IsFormat_DOLT(types.Format_Default) {
-		for _, test := range Dolt1DiffSystemTableScripts {
-			func() {
-				h = h.NewHarness(t)
-				defer h.Close()
-				h.Setup(setup.MydbData)
-				enginetest.TestScript(t, h, test)
-			}()
-		}
-	}
-}
-
 func TestDiffSystemTablePrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltDiffSystemTableTestsPrepared(t, h)
-}
-
-func RunDoltDiffSystemTableTestsPrepared(t *testing.T, h DoltEnginetestHarness) {
-	if !types.IsFormat_DOLT(types.Format_Default) {
-		t.Skip("only new format support system table indexing")
-	}
-
-	for _, test := range DiffSystemTableScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			h = h.NewHarness(t)
-			defer h.Close()
-			h.Setup(setup.MydbData)
-			enginetest.TestScriptPrepared(t, h, test)
-		})
-	}
-
-	if types.IsFormat_DOLT(types.Format_Default) {
-		for _, test := range Dolt1DiffSystemTableScripts {
-			func() {
-				h = h.NewHarness(t)
-				defer h.Close()
-				h.Setup(setup.MydbData)
-				enginetest.TestScriptPrepared(t, h, test)
-			}()
-		}
-	}
 }
 
 func TestSchemaDiffTableFunction(t *testing.T) {
@@ -2867,31 +1557,9 @@ func TestSchemaDiffTableFunction(t *testing.T) {
 	RunSchemaDiffTableFunctionTests(t, harness)
 }
 
-func RunSchemaDiffTableFunctionTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range SchemaDiffTableFunctionScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			enginetest.TestScript(t, harness, test)
-		})
-	}
-}
-
 func TestSchemaDiffTableFunctionPrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunSchemaDiffTableFunctionTestsPrepared(t, harness)
-}
-
-func RunSchemaDiffTableFunctionTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range SchemaDiffTableFunctionScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			enginetest.TestScriptPrepared(t, harness, test)
-		})
-	}
 }
 
 func TestDoltDatabaseCollationDiffs(t *testing.T) {
@@ -2899,46 +1567,9 @@ func TestDoltDatabaseCollationDiffs(t *testing.T) {
 	RunDoltDatabaseCollationDiffsTests(t, harness)
 }
 
-func RunDoltDatabaseCollationDiffsTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range DoltDatabaseCollationScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			enginetest.TestScriptPrepared(t, harness, test)
-		})
-	}
-}
-
 func TestQueryDiff(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunQueryDiffTests(t, harness)
-}
-
-func RunQueryDiffTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range QueryDiffTableScriptTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			harness.Setup(setup.MydbData)
-			enginetest.TestScript(t, harness, test)
-		})
-	}
-}
-
-func mustNewEngine(t *testing.T, h enginetest.Harness) enginetest.QueryEngine {
-	e, err := h.NewEngine(t)
-	if err != nil {
-		require.NoError(t, err)
-	}
-	return e
-}
-
-var biasedCosters = []memo.Coster{
-	memo.NewInnerBiasedCoster(),
-	memo.NewLookupBiasedCoster(),
-	memo.NewHashBiasedCoster(),
-	memo.NewMergeBiasedCoster(),
 }
 
 func TestSystemTableIndexes(t *testing.T) {
@@ -2946,84 +1577,9 @@ func TestSystemTableIndexes(t *testing.T) {
 	RunSystemTableIndexesTests(t, harness)
 }
 
-func RunSystemTableIndexesTests(t *testing.T, harness DoltEnginetestHarness) {
-	if !types.IsFormat_DOLT(types.Format_Default) {
-		t.Skip("only new format support system table indexing")
-	}
-
-	for _, stt := range SystemTableIndexTests {
-		harness = harness.NewHarness(t).WithParallelism(2)
-		defer harness.Close()
-		harness.SkipSetupCommit()
-		e := mustNewEngine(t, harness)
-		defer e.Close()
-		e.EngineAnalyzer().Coster = memo.NewMergeBiasedCoster()
-
-		ctx := enginetest.NewContext(harness)
-		for _, q := range stt.setup {
-			enginetest.RunQueryWithContext(t, e, harness, ctx, q)
-		}
-
-		for i, c := range []string{"inner", "lookup", "hash", "merge"} {
-			e.EngineAnalyzer().Coster = biasedCosters[i]
-			for _, tt := range stt.queries {
-				if tt.query == "select count(*) from dolt_blame_xy" && c == "inner" {
-					// todo we either need join hints to work inside the blame view
-					// and force the window relation to be primary, or we need the
-					// blame view's timestamp columns to be specific enough to not
-					// overlap during testing.
-					t.Skip("the blame table is unstable as secondary table in join with exchange node")
-				}
-				t.Run(fmt.Sprintf("%s(%s): %s", stt.name, c, tt.query), func(t *testing.T) {
-					if tt.skip {
-						t.Skip()
-					}
-
-					ctx = ctx.WithQuery(tt.query)
-					if tt.exp != nil {
-						enginetest.TestQueryWithContext(t, ctx, e, harness, tt.query, tt.exp, nil, nil)
-					}
-				})
-			}
-		}
-	}
-}
-
 func TestSystemTableIndexesPrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunSystemTableIndexesTestsPrepared(t, harness)
-}
-
-func RunSystemTableIndexesTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	if !types.IsFormat_DOLT(types.Format_Default) {
-		t.Skip("only new format support system table indexing")
-	}
-
-	for _, stt := range SystemTableIndexTests {
-		harness = harness.NewHarness(t).WithParallelism(2)
-		defer harness.Close()
-		harness.SkipSetupCommit()
-		e := mustNewEngine(t, harness)
-		defer e.Close()
-
-		ctx := enginetest.NewContext(harness)
-		for _, q := range stt.setup {
-			enginetest.RunQueryWithContext(t, e, harness, ctx, q)
-		}
-
-		for _, tt := range stt.queries {
-			t.Run(fmt.Sprintf("%s: %s", stt.name, tt.query), func(t *testing.T) {
-				if tt.skip {
-					t.Skip()
-				}
-
-				ctx = ctx.WithQuery(tt.query)
-				if tt.exp != nil {
-					enginetest.TestPreparedQueryWithContext(t, ctx, e, harness, tt.query, tt.exp, nil, nil, false)
-				}
-			})
-		}
-	}
 }
 
 func TestSystemTableFunctionIndexes(t *testing.T) {
@@ -3031,29 +1587,9 @@ func TestSystemTableFunctionIndexes(t *testing.T) {
 	RunSystemTableFunctionIndexesTests(t, harness)
 }
 
-func RunSystemTableFunctionIndexesTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range SystemTableFunctionIndexTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			harness.Setup(setup.MydbData)
-			enginetest.TestScript(t, harness, test)
-		})
-	}
-}
-
 func TestSystemTableFunctionIndexesPrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunSystemTableFunctionIndexesTestsPrepared(t, harness)
-}
-
-func RunSystemTableFunctionIndexesTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	for _, test := range SystemTableFunctionIndexTests {
-		t.Run(test.Name, func(t *testing.T) {
-			harness = harness.NewHarness(t)
-			harness.Setup(setup.MydbData)
-			enginetest.TestScriptPrepared(t, harness, test)
-		})
-	}
 }
 
 func TestReadOnlyDatabases(t *testing.T) {
@@ -3071,13 +1607,6 @@ func TestAddDropPks(t *testing.T) {
 func TestAddAutoIncrementColumn(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunAddAutoIncrementColumnTests(t, h)
-}
-
-func RunAddAutoIncrementColumnTests(t *testing.T, h DoltEnginetestHarness) {
-	defer h.Close()
-	for _, script := range queries.AlterTableAddAutoIncrementScripts {
-		enginetest.TestScript(t, h, script)
-	}
 }
 
 func TestNullRanges(t *testing.T) {
@@ -3115,25 +1644,9 @@ func TestDoltCherryPick(t *testing.T) {
 	RunDoltCherryPickTests(t, harness)
 }
 
-func RunDoltCherryPickTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, script := range DoltCherryPickTests {
-		harness = harness.NewHarness(t)
-		enginetest.TestScript(t, harness, script)
-		harness.Close()
-	}
-}
-
 func TestDoltCherryPickPrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunDoltCherryPickTestsPrepared(t, harness)
-}
-
-func RunDoltCherryPickTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	for _, script := range DoltCherryPickTests {
-		harness = harness.NewHarness(t)
-		enginetest.TestScriptPrepared(t, harness, script)
-		harness.Close()
-	}
 }
 
 func TestDoltCommit(t *testing.T) {
@@ -3141,23 +1654,9 @@ func TestDoltCommit(t *testing.T) {
 	RunDoltCommitTests(t, harness)
 }
 
-func RunDoltCommitTests(t *testing.T, harness DoltEnginetestHarness) {
-	defer harness.Close()
-	for _, script := range DoltCommitTests {
-		enginetest.TestScript(t, harness, script)
-	}
-}
-
 func TestDoltCommitPrepared(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunDoltCommitTestsPrepared(t, harness)
-}
-
-func RunDoltCommitTestsPrepared(t *testing.T, harness DoltEnginetestHarness) {
-	defer harness.Close()
-	for _, script := range DoltCommitTests {
-		enginetest.TestScriptPrepared(t, harness, script)
-	}
 }
 
 func TestQueriesPrepared(t *testing.T) {
@@ -3171,16 +1670,6 @@ func TestStatsHistograms(t *testing.T) {
 	RunStatsHistogramTests(t, h)
 }
 
-func RunStatsHistogramTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltHistogramTests {
-		func() {
-			h = h.NewHarness(t).WithConfigureStats(true)
-			defer h.Close()
-			enginetest.TestScript(t, h, script)
-		}()
-	}
-}
-
 // TestStatsIO force a provider reload in-between setup and assertions that
 // forces a round trip of the statistics table before inspecting values.
 func TestStatsIO(t *testing.T) {
@@ -3188,33 +1677,9 @@ func TestStatsIO(t *testing.T) {
 	RunStatsIOTests(t, h)
 }
 
-func RunStatsIOTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range append(DoltStatsIOTests, DoltHistogramTests...) {
-		func() {
-			h = h.NewHarness(t).WithConfigureStats(true)
-			defer h.Close()
-			e := mustNewEngine(t, h)
-			if enginetest.IsServerEngine(e) {
-				return
-			}
-			defer e.Close()
-			TestProviderReloadScriptWithEngine(t, e, h, script)
-		}()
-	}
-}
-
 func TestJoinStats(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunJoinStatsTests(t, h)
-}
-
-// these are sensitive to cardinality estimates,
-// particularly the join-filter tests that trade-off
-// smallest table first vs smallest join first
-func RunJoinStatsTests(t *testing.T, h DoltEnginetestHarness) {
-	defer h.Close()
-	h = h.WithConfigureStats(true)
-	enginetest.TestJoinStats(t, h)
 }
 
 func TestStatisticIndexes(t *testing.T) {
@@ -3236,35 +1701,9 @@ func TestPreparedStatistics(t *testing.T) {
 	RunPreparedStatisticsTests(t, h)
 }
 
-func RunPreparedStatisticsTests(t *testing.T, h DoltEnginetestHarness) {
-	for _, script := range DoltHistogramTests {
-		func() {
-			h := h.NewHarness(t).WithConfigureStats(true)
-			defer h.Close()
-			enginetest.TestScriptPrepared(t, h, script)
-		}()
-	}
-}
-
 func TestVersionedQueriesPrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunVersionedQueriesPreparedTests(t, h)
-}
-
-func RunVersionedQueriesPreparedTests(t *testing.T, h DoltEnginetestHarness) {
-	defer h.Close()
-	h.Setup(setup.MydbData, []setup.SetupScript{VersionedQuerySetup, VersionedQueryViews})
-
-	e, err := h.NewEngine(t)
-	require.NoError(t, err)
-
-	for _, tt := range queries.VersionedQueries {
-		enginetest.TestPreparedQueryWithEngine(t, h, e, tt)
-	}
-
-	for _, tt := range queries.VersionedScripts {
-		enginetest.TestScriptWithEnginePrepared(t, e, h, tt)
-	}
 }
 
 func TestInfoSchemaPrepared(t *testing.T) {
@@ -3438,232 +1877,14 @@ func TestAddDropPrimaryKeys(t *testing.T) {
 	RunAddDropPrimaryKeysTests(t, harness)
 }
 
-func RunAddDropPrimaryKeysTests(t *testing.T, harness DoltEnginetestHarness) {
-	t.Run("adding and dropping primary keys does not result in duplicate NOT NULL constraints", func(t *testing.T) {
-		harness = harness.NewHarness(t)
-		defer harness.Close()
-		addPkScript := queries.ScriptTest{
-			Name: "add primary keys",
-			SetUpScript: []string{
-				"create table test (id int not null, c1 int);",
-				"create index c1_idx on test(c1)",
-				"insert into test values (1,1),(2,2)",
-				"ALTER TABLE test ADD PRIMARY KEY(id)",
-				"ALTER TABLE test DROP PRIMARY KEY",
-				"ALTER TABLE test ADD PRIMARY KEY(id)",
-				"ALTER TABLE test DROP PRIMARY KEY",
-				"ALTER TABLE test ADD PRIMARY KEY(id)",
-				"ALTER TABLE test DROP PRIMARY KEY",
-				"ALTER TABLE test ADD PRIMARY KEY(id)",
-			},
-			Assertions: []queries.ScriptTestAssertion{
-				{
-					Query: "show create table test",
-					Expected: []sql.Row{
-						{"test", "CREATE TABLE `test` (\n" +
-								"  `id` int NOT NULL,\n" +
-								"  `c1` int,\n" +
-								"  PRIMARY KEY (`id`),\n" +
-								"  KEY `c1_idx` (`c1`)\n" +
-								") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_bin"},
-					},
-				},
-			},
-		}
-
-		enginetest.TestScript(t, harness, addPkScript)
-
-		// make sure there is only one NOT NULL constraint after all those mutations
-		ctx := sql.NewContext(context.Background(), sql.WithSession(harness.Session()))
-		ws, err := harness.Session().WorkingSet(ctx, "mydb")
-		require.NoError(t, err)
-
-		table, ok, err := ws.WorkingRoot().GetTable(ctx, doltdb.TableName{Name: "test"})
-		require.NoError(t, err)
-		require.True(t, ok)
-
-		sch, err := table.GetSchema(ctx)
-		for _, col := range sch.GetAllCols().GetColumns() {
-			count := 0
-			for _, cc := range col.Constraints {
-				if cc.GetConstraintType() == schema.NotNullConstraintType {
-					count++
-				}
-			}
-			require.Less(t, count, 2)
-		}
-	})
-
-	t.Run("Add primary key to table with index", func(t *testing.T) {
-		harness := harness.NewHarness(t)
-		defer harness.Close()
-		script := queries.ScriptTest{
-			Name: "add primary keys to table with index",
-			SetUpScript: []string{
-				"create table test (id int not null, c1 int);",
-				"create index c1_idx on test(c1)",
-				"insert into test values (1,1),(2,2)",
-				"ALTER TABLE test ADD constraint test_check CHECK (c1 > 0)",
-				"ALTER TABLE test ADD PRIMARY KEY(id)",
-			},
-			Assertions: []queries.ScriptTestAssertion{
-				{
-					Query: "show create table test",
-					Expected: []sql.Row{
-						{"test", "CREATE TABLE `test` (\n" +
-								"  `id` int NOT NULL,\n" +
-								"  `c1` int,\n" +
-								"  PRIMARY KEY (`id`),\n" +
-								"  KEY `c1_idx` (`c1`),\n" +
-								"  CONSTRAINT `test_check` CHECK ((`c1` > 0))\n" +
-								") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_bin"},
-					},
-				},
-				{
-					Query: "select * from test order by id",
-					Expected: []sql.Row{
-						{1, 1},
-						{2, 2},
-					},
-				},
-			},
-		}
-		enginetest.TestScript(t, harness, script)
-
-		ctx := sql.NewContext(context.Background(), sql.WithSession(harness.Session()))
-		ws, err := harness.Session().WorkingSet(ctx, "mydb")
-		require.NoError(t, err)
-
-		table, ok, err := ws.WorkingRoot().GetTable(ctx, doltdb.TableName{Name: "test"})
-		require.NoError(t, err)
-		require.True(t, ok)
-
-		// Assert the new index map is not empty
-		newRows, err := table.GetIndexRowData(ctx, "c1_idx")
-		require.NoError(t, err)
-		empty, err := newRows.Empty()
-		require.NoError(t, err)
-		assert.False(t, empty)
-		count, err := newRows.Count()
-		require.NoError(t, err)
-		assert.Equal(t, count, uint64(2))
-	})
-
-	t.Run("Add primary key when one more cells contain NULL", func(t *testing.T) {
-		harness := harness.NewHarness(t)
-		defer harness.Close()
-		script := queries.ScriptTest{
-			Name: "Add primary key when one more cells contain NULL",
-			SetUpScript: []string{
-				"create table test (id int not null, c1 int);",
-				"create index c1_idx on test(c1)",
-				"insert into test values (1,1),(2,2)",
-				"ALTER TABLE test ADD PRIMARY KEY (c1)",
-				"ALTER TABLE test ADD COLUMN (c2 INT NULL)",
-				"ALTER TABLE test DROP PRIMARY KEY",
-			},
-			Assertions: []queries.ScriptTestAssertion{
-				{
-					Query:       "ALTER TABLE test ADD PRIMARY KEY (id, c1, c2)",
-					ExpectedErr: sql.ErrInsertIntoNonNullableProvidedNull,
-				},
-			},
-		}
-		enginetest.TestScript(t, harness, script)
-	})
-
-	t.Run("Drop primary key from table with index", func(t *testing.T) {
-		harness := harness.NewHarness(t)
-		defer harness.Close()
-		script := queries.ScriptTest{
-			Name: "Drop primary key from table with index",
-			SetUpScript: []string{
-				"create table test (id int not null primary key, c1 int);",
-				"create index c1_idx on test(c1)",
-				"insert into test values (1,1),(2,2)",
-				"ALTER TABLE test DROP PRIMARY KEY",
-			},
-			Assertions: []queries.ScriptTestAssertion{
-				{
-					Query: "show create table test",
-					Expected: []sql.Row{
-						{"test", "CREATE TABLE `test` (\n" +
-								"  `id` int NOT NULL,\n" +
-								"  `c1` int,\n" +
-								"  KEY `c1_idx` (`c1`)\n" +
-								") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_bin"},
-					},
-				},
-				{
-					Query: "select * from test order by id",
-					Expected: []sql.Row{
-						{1, 1},
-						{2, 2},
-					},
-				},
-			},
-		}
-
-		enginetest.TestScript(t, harness, script)
-
-		ctx := sql.NewContext(context.Background(), sql.WithSession(harness.Session()))
-		ws, err := harness.Session().WorkingSet(ctx, "mydb")
-		require.NoError(t, err)
-
-		table, ok, err := ws.WorkingRoot().GetTable(ctx, doltdb.TableName{Name: "test"})
-		require.NoError(t, err)
-		require.True(t, ok)
-
-		// Assert the index map is not empty
-		newIdx, err := table.GetIndexRowData(ctx, "c1_idx")
-		assert.NoError(t, err)
-		empty, err := newIdx.Empty()
-		require.NoError(t, err)
-		assert.False(t, empty)
-		count, err := newIdx.Count()
-		require.NoError(t, err)
-		assert.Equal(t, count, uint64(2))
-	})
-}
-
 func TestDoltVerifyConstraints(t *testing.T) {
 	harness := newDoltEnginetestHarness(t)
 	RunDoltVerifyConstraintsTests(t, harness)
 }
 
-func RunDoltVerifyConstraintsTests(t *testing.T, harness DoltEnginetestHarness) {
-	for _, script := range DoltVerifyConstraintsTestScripts {
-		func() {
-			harness = harness.NewHarness(t)
-			defer harness.Close()
-			enginetest.TestScript(t, harness, script)
-		}()
-	}
-}
-
 func TestDoltStorageFormat(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 	RunDoltStorageFormatTests(t, h)
-}
-
-func RunDoltStorageFormatTests(t *testing.T, h DoltEnginetestHarness) {
-	var expectedFormatString string
-	if types.IsFormat_DOLT(types.Format_Default) {
-		expectedFormatString = "NEW ( __DOLT__ )"
-	} else {
-		expectedFormatString = fmt.Sprintf("OLD ( %s )", types.Format_Default.VersionString())
-	}
-	script := queries.ScriptTest{
-		Name: "dolt storage format function works",
-		Assertions: []queries.ScriptTestAssertion{
-			{
-				Query:    "select dolt_storage_format()",
-				Expected: []sql.Row{{expectedFormatString}},
-			},
-		},
-	}
-	defer h.Close()
-	enginetest.TestScript(t, h, script)
 }
 
 func TestDoltStorageFormatPrepared(t *testing.T) {
@@ -3684,57 +1905,12 @@ func TestThreeWayMergeWithSchemaChangeScripts(t *testing.T) {
 	RunThreeWayMergeWithSchemaChangeScripts(t, h)
 }
 
-func RunThreeWayMergeWithSchemaChangeScripts(t *testing.T, h DoltEnginetestHarness) {
-	skipOldFormat(t)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsBasicCases, "basic cases", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsForDataConflicts, "data conflicts", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsCollations, "collation changes", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsConstraints, "constraint changes", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsSchemaConflicts, "schema conflicts", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsGeneratedColumns, "generated columns", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsForJsonConflicts, "json merge", false)
-
-	// Run non-symmetric schema merge tests in just one direction
-	t.Run("type changes", func(t *testing.T) {
-		for _, script := range SchemaChangeTestsTypeChanges {
-			// run in a func() so we can cleanly defer closing the harness
-			func() {
-				h := h.NewHarness(t)
-				defer h.Close()
-				enginetest.TestScript(t, h, convertMergeScriptTest(script, false))
-			}()
-		}
-	})
-}
-
 func TestThreeWayMergeWithSchemaChangeScriptsPrepared(t *testing.T) {
 	h := newDoltEnginetestHarness(t)
 
 	RunThreeWayMergeWithSchemaChangeScriptsPrepared(t, h)
 }
 
-func RunThreeWayMergeWithSchemaChangeScriptsPrepared(t *testing.T, h DoltEnginetestHarness) {
-	skipOldFormat(t)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsBasicCases, "basic cases", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsForDataConflicts, "data conflicts", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsCollations, "collation changes", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsConstraints, "constraint changes", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsSchemaConflicts, "schema conflicts", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsGeneratedColumns, "generated columns", false)
-	runMergeScriptTestsInBothDirections(t, SchemaChangeTestsForJsonConflicts, "json merge", false)
-
-	// Run non-symmetric schema merge tests in just one direction
-	t.Run("type changes", func(t *testing.T) {
-		for _, script := range SchemaChangeTestsTypeChanges {
-			// run in a func() so we can cleanly defer closing the harness
-			func() {
-				h := h.NewHarness(t)
-				defer h.Close()
-				enginetest.TestScriptPrepared(t, h, convertMergeScriptTest(script, false))
-			}()
-		}
-	})
-}
 // If CREATE DATABASE has an error within the DatabaseProvider, it should not
 // leave behind intermediate filesystem state.
 func TestCreateDatabaseErrorCleansUp(t *testing.T) {
@@ -3871,66 +2047,5 @@ func TestStatsAutoRefreshConcurrency(t *testing.T) {
 		}()
 
 		wg.Wait()
-	}
-}
-
-// runMergeScriptTestsInBothDirections creates a new test run, named |name|, and runs the specified merge |tests|
-// in both directions (right to left merge, and left to right merge). If
-// |runAsPrepared| is true then the test scripts will be run using the prepared
-// statement test code.
-func runMergeScriptTestsInBothDirections(t *testing.T, tests []MergeScriptTest, name string, runAsPrepared bool) {
-	t.Run(name, func(t *testing.T) {
-		t.Run("right to left merges", func(t *testing.T) {
-			for _, script := range tests {
-				// run in a func() so we can cleanly defer closing the harness
-				func() {
-					h := newDoltHarness(t)
-					defer h.Close()
-					if runAsPrepared {
-						enginetest.TestScriptPrepared(t, h, convertMergeScriptTest(script, false))
-					} else {
-						enginetest.TestScript(t, h, convertMergeScriptTest(script, false))
-					}
-				}()
-			}
-		})
-		t.Run("left to right merges", func(t *testing.T) {
-			for _, script := range tests {
-				func() {
-					h := newDoltHarness(t)
-					defer h.Close()
-					if runAsPrepared {
-						enginetest.TestScriptPrepared(t, h, convertMergeScriptTest(script, true))
-					} else {
-						enginetest.TestScript(t, h, convertMergeScriptTest(script, true))
-					}
-				}()
-			}
-		})
-	})
-}
-
-var newFormatSkippedScripts = []string{
-	// Different query plans
-	"Partial indexes are used and return the expected result",
-	"Multiple indexes on the same columns in a different order",
-}
-
-func skipOldFormat(t *testing.T) {
-	if !types.IsFormat_DOLT(types.Format_Default) {
-		t.Skip()
-	}
-}
-
-func skipPreparedTests(t *testing.T) {
-	if skipPrepared {
-		t.Skip("skip prepared")
-	}
-}
-
-func newSessionBuilder(harness *DoltHarness) server.SessionBuilder {
-	return func(ctx context.Context, conn *mysql.Conn, host string) (sql.Session, error) {
-		newCtx := harness.NewSession()
-		return newCtx.Session, nil
 	}
 }
